// src/types/Synergetic/iVStaff.ts
type iVStaff = {
  ActiveFlag: boolean;
  CommunityGUID: string;
  Doctor2Address: string;
  Doctor2Name: string;
  Doctor2Phone: string;
  Doctor2PhoneAlt: string;
  Doctor2Specialisation: string;
  DoctorAddress: string;
  DoctorName: string;
  DoctorPhone: string;
  DoctorPhoneAlt: string;
  DoctorSpecialisation: string;
  Emergency2Address: string;
  Emergency2Name: string;
  Emergency2Phone: string;
  Emergency2PhoneAlt: string;
  Emergency2Relat: string;
  EmergencyAddress: string;
  EmergencyName: string;
  EmergencyPhone: string;
  EmergencyPhoneAlt: string;
  EmergencyRelat: string;
  FirstTeachingYear: boolean;
  LeavingDate: Date | null;
  LicenseNumber: string;
  MailingAddress: string;
  MailingName: string;
  MedicalAlert: string;
  MedicalAlertFlag: boolean;
  PoliceCheckReferenceNumber: string;
  PoliceCheckedDate: Date | null;
  PoliceCheckedFlag: boolean;
  ProbationFromDate: Date | null;
  ProbationToDate: Date | null;
  ReferencesCheckedFlag: boolean;
  SchoolStaffCode: string;
  StaffAddressComma: string;
  StaffAddressDPID: string;
  StaffAddressDPIDBarcode: string;
  StaffAddressFull: string;
  StaffAddressPAFVersion: string;
  StaffAddressSortPlanNumber: string;
  StaffBirthDate: Date;
  StaffBuilding: string;
  StaffCampus: string;
  StaffCampusDescription: string;
  StaffCategory: string;
  StaffCategoryDescription: string;
  StaffCategoryType: string;
  StaffDefaultMobilePhone: string;
  StaffDepartment: string;
  StaffExtension: string;
  StaffForm: string;
  StaffGender: string;
  StaffGenderSynergyMeaning: string;
  StaffGiven1: string;
  StaffGiven2: string;
  StaffHouse: string;
  StaffID: number;
  StaffLegalFullName: string;
  StaffMailNamePrimary: string;
  StaffMobilePhone: string;
  StaffNameExternal: string;
  StaffNameInternal: string;
  StaffOccupEmail: string;
  StaffOccupMobilePhone: string;
  StaffOccupPhone: string;
  StaffPhone: string;
  StaffPreferred: string;
  StaffPreferredFormal: string;
  StaffPreferredName: string;
  StaffRegSchoolsBoardCode: string;
  StaffRegistrationExpiryDate: Date | null;
  StaffRegistrationNumber: string;
  StaffRegistrationType: string;
  StaffRoom: string;
  StaffSubstituteAvailableFlag: boolean;
  StaffSurname: string;
  StaffTitle: string;
  StaffVoiceMail: string;
  StaffYearLevel: string | null;
  StartDate: Date;
  SubSchool: string | null;
  ValidAddressFlag: boolean;
  WebSite: string;
  WorkWithChildrenExpiryDate: Date;
  WorkWithChildrenRegistrationNumber: string;
  WorkWithChildrenStatus: string;
  WorkWithChildrenStatusDescription: string;
  WorkWithCildrenRegistrationNumber: string;
};

export default iVStaff;